fix(#711): move rebac domain model from core/ to rebac/domain.py
Move core/rebac.py (pure domain models: Entity, ReBACTuple, NamespaceConfig,
CheckCacheEntry, ChangelogEntry) from kernel to rebac/ brick where it belongs.
The kernel only needs a check(subject, permission, object) protocol, not the
full tuple/namespace domain model. Updated all 60 import sites across 28 files
(src, tests, docs).
