+++ src/nexus/rebac/batch/bulk_checker.py
@@ -22,8 +22,8 @@ from typing import TYPE_CHECKING, Any
 
 from sqlalchemy.exc import OperationalError
 
-from nexus.core.rebac import Entity
 from nexus.rebac.cross_zone import CROSS_ZONE_ALLOWED_RELATIONS
+from nexus.rebac.domain import Entity
 from nexus.rebac.types import ConsistencyLevel
 
 if TYPE_CHECKING:
@@ -32,8 +32,8 @@ if TYPE_CHECKING:
 
     from sqlalchemy.engine import Engine
 
-    from nexus.core.rebac import NamespaceConfig
     from nexus.rebac.cache.tiger import TigerCache
+    from nexus.rebac.domain import NamespaceConfig
 
 logger = logging.getLogger(__name__)
