+++ src/nexus/rebac/graph/traversal.py
@@ -16,12 +16,12 @@ import logging
 from datetime import UTC, datetime
 from typing import TYPE_CHECKING, Any
 
-from nexus.core.rebac import WILDCARD_SUBJECT, Entity
+from nexus.rebac.domain import WILDCARD_SUBJECT, Entity
 
 if TYPE_CHECKING:
     from collections.abc import Callable
 
-    from nexus.core.rebac import NamespaceConfig
+    from nexus.rebac.domain import NamespaceConfig
     from nexus.rebac.tuples.repository import TupleRepository
 
 logger = logging.getLogger(__name__)
