+++ src/nexus/rebac/manager.py
@@ -28,10 +28,6 @@ from typing import TYPE_CHECKING, Any, cast
 
 from sqlalchemy.exc import OperationalError, ProgrammingError
 
-from nexus.core.rebac import (
-    Entity,
-    NamespaceConfig,
-)
 from nexus.rebac.batch.bulk_checker import BulkPermissionChecker
 from nexus.rebac.cache.result_cache import ReBACPermissionCache
 from nexus.rebac.cache.tiger.facade import TigerFacade
@@ -42,6 +38,10 @@ from nexus.rebac.consistency.revision import (
 from nexus.rebac.consistency.zone_manager import ZoneManager
 from nexus.rebac.cross_zone import CROSS_ZONE_ALLOWED_RELATIONS
 from nexus.rebac.directory.expander import DirectoryExpander
+from nexus.rebac.domain import (
+    Entity,
+    NamespaceConfig,
+)
 from nexus.rebac.graph.bulk_evaluator import (
     check_direct_relation as _check_direct_relation_in_graph,
 )
@@ -553,7 +553,7 @@ class ReBACManager:
         subject_type, subject_id = subject
         object_type, object_id = object
 
-        from nexus.core.rebac import WILDCARD_SUBJECT
+        from nexus.rebac.domain import WILDCARD_SUBJECT
 
         with self._connection(readonly=True) as conn:
             cursor = self._create_cursor(conn)
@@ -1189,7 +1189,7 @@ class ReBACManager:
         Returns:
             List of wildcard tuples from other zones
         """
-        from nexus.core.rebac import WILDCARD_SUBJECT
+        from nexus.rebac.domain import WILDCARD_SUBJECT
 
         with self._connection(readonly=True) as conn:
             cursor = self._create_cursor(conn)
@@ -3679,7 +3679,7 @@ class ReBACManager:
         import json
         from datetime import datetime
 
-        from nexus.core.rebac import NamespaceConfig
+        from nexus.rebac.domain import NamespaceConfig
 
         with self._connection(readonly=True) as conn:
             cursor = self._create_cursor(conn)
@@ -3933,7 +3933,7 @@ class ReBACManager:
         import uuid
         from datetime import UTC, datetime
 
-        from nexus.core.rebac import Entity
+        from nexus.rebac.domain import Entity
 
         logger = logging.getLogger(__name__)
 
@@ -4219,7 +4219,7 @@ class ReBACManager:
         """
         from datetime import UTC, datetime
 
-        from nexus.core.rebac import Entity
+        from nexus.rebac.domain import Entity
 
         with self._connection() as conn:
             cursor = self._create_cursor(conn)
@@ -4354,7 +4354,7 @@ class ReBACManager:
         import logging
         from datetime import UTC, datetime
 
-        from nexus.core.rebac import Entity
+        from nexus.rebac.domain import Entity
 
         logger = logging.getLogger(__name__)
 
@@ -4794,7 +4794,7 @@ class ReBACManager:
         """
         import logging
 
-        from nexus.core.rebac import Entity
+        from nexus.rebac.domain import Entity
 
         logger = logging.getLogger(__name__)
 
@@ -4958,7 +4958,7 @@ class ReBACManager:
             ... ])
             >>> # Returns: [True, False, True]
         """
-        from nexus.core.rebac import Entity
+        from nexus.rebac.domain import Entity
 
         if not checks:
             return []
@@ -5030,7 +5030,7 @@ class ReBACManager:
         """
         import logging
 
-        from nexus.core.rebac import Entity
+        from nexus.rebac.domain import Entity
 
         logger = logging.getLogger(__name__)
 
@@ -5116,7 +5116,7 @@ class ReBACManager:
         """Compute uncached checks using Python (original implementation)."""
         import time as time_module
 
-        from nexus.core.rebac import Entity
+        from nexus.rebac.domain import Entity
 
         for i, (subject, permission, obj) in uncached_checks:
             subject_entity = Entity(subject[0], subject[1])
@@ -5285,7 +5285,7 @@ class ReBACManager:
         import uuid
         from datetime import UTC, datetime
 
-        from nexus.core.rebac import Entity
+        from nexus.rebac.domain import Entity
 
         # Generate request ID and timestamp
         request_id = f"req_{uuid.uuid4().hex[:12]}"
@@ -5593,7 +5593,7 @@ class ReBACManager:
         """
         import logging
 
-        from nexus.core.rebac import Entity
+        from nexus.rebac.domain import Entity
 
         logger = logging.getLogger(__name__)
 
@@ -6116,7 +6116,7 @@ class ReBACManager:
         """
         from datetime import UTC, datetime
 
-        from nexus.core.rebac import Entity
+        from nexus.rebac.domain import Entity
 
         with self._connection() as conn:
             cursor = self._create_cursor(conn)
