+++ src/nexus/services/permissions/utils/fast.py
@@ -18,8 +18,8 @@ import logging
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
-    from nexus.core.rebac import Entity
-    from nexus.core.rebac import NamespaceConfig as ReBACNamespaceConfig
+    from nexus.rebac.domain import Entity
+    from nexus.rebac.domain import NamespaceConfig as ReBACNamespaceConfig
 
 
 # Internal type for namespace config dict (not the NamespaceConfig class)
@@ -194,7 +194,7 @@ def _check_permissions_bulk_python(
     This is a simplified implementation. For production, this should delegate
     to the existing ReBACManager._compute_permission logic.
     """
-    from nexus.core.rebac import Entity, NamespaceConfig
+    from nexus.rebac.domain import Entity, NamespaceConfig
 
     # Convert namespace configs to proper format
     namespaces: dict[str, ReBACNamespaceConfig] = {}
