+++ tests/unit/core/test_rebac.py
@@ -17,7 +17,7 @@ import pytest
 from freezegun import freeze_time
 from sqlalchemy import create_engine
 
-from nexus.core.rebac import Entity, NamespaceConfig
+from nexus.rebac.domain import Entity, NamespaceConfig
 from nexus.rebac.manager import ReBACManager
 from nexus.storage.models import Base
