+++ tests/unit/services/permissions/graph/test_traversal_edge_cases.py
@@ -28,7 +28,7 @@ import pytest
 from freezegun import freeze_time
 from sqlalchemy import create_engine
 
-from nexus.core.rebac import NamespaceConfig
+from nexus.rebac.domain import NamespaceConfig
 from nexus.rebac.manager import ReBACManager
 from nexus.storage.models import Base
